Starting to update elements with style changes
